requirements.txt
langchain
openai
python-dotenv
requests
unittest2
